+++ Citadel/src/utopia/citadel/database/model/user/UserDeviceModel.scala
@@ -2,7 +2,7 @@ package utopia.citadel.database.model.user
 
 import java.time.Instant
 import utopia.citadel.database.Tables
-import utopia.citadel.database.model.DeprecatableAfter
+import utopia.vault.nosql.storable.deprecation.DeprecatableAfter
 import utopia.flow.generic.ValueConversions._
 import utopia.vault.database.Connection
 import utopia.vault.model.immutable.Storable